Adding Version 0,1 to CaloTPGTranscoderULUT
Version 0 uses the old segmentation (<= 2012), while Version 1 uses the
full segmented HF (> 2012).

This has necessitated fixing calls to HTvalid and getOutputLUTId in
other pieces of code. For cases where a HcalTrigTowerDetId was not
available (and hence we could not use its internal version variable), we
have hard coded the version to 0.

Remove unused hcaletValue functions

Removed from CaloTPGTranscoderULUT and CaloTPGTranscoder.